+++ langs/swift/Sources/VMx/Builders/CompositeVMBuilder.swift
@@ -32,6 +32,12 @@ public struct CompositeVMBuilder<Child: ComponentVMBase> {
     ) -> CompositeVMBuilder<Child> {
         var c = self; c._hub = hub; c._dispatcher = dispatcher; return c
     }
+    func _optionHub(_ hub: MessageHubProtocol) -> CompositeVMBuilder<Child> {
+        var c = self; c._hub = hub; return c
+    }
+    func _optionDispatcher(_ dispatcher: Dispatcher) -> CompositeVMBuilder<Child> {
+        var c = self; c._dispatcher = dispatcher; return c
+    }
     public func children(
         _ factory: @escaping () -> [Child]
     ) -> CompositeVMBuilder<Child> {